// packages/shared/src/types/nodes.ts
export interface ImageData extends NodeData {
  src?: string | null;
  width?: number | null;
  height?: number | null;
  fit?: string;
  color?: string | number;
  pivot?: string;
  lit?: boolean;
  doubleside?: boolean;
  castShadow?: boolean;
  receiveShadow?: boolean;
  emissive?: string | number;
  emissiveIntensity?: number;
  opacity?: number;
  transparent?: boolean;
  alphaTest?: number;
}